+++ src/lib/vulkan/shaders/wire.mesh.geom
@@ -14,10 +14,10 @@ layout (binding = 1) uniform UBO
 } ubo;
 
 layout (location = 0) in vec3 inNormal[]; 
-layout (location = 1) in vec4 inId[]; 
+layout (location = 1) in ivec4 inId[]; 
 
 layout (location = 0) out vec3 outColor;
-layout (location = 1) out vec4 outId;
+layout (location = 1) out ivec4 outId;
 
 void main() 
 {
